+++ docs/content/concepts/ops-jobs-graphs/ops.mdx
@@ -177,3 +177,9 @@ def my_op_factory(
 
     return my_inner_op
 ```
+
+## See it in action
+For more examples of ops, check out the following in our [Hacker News example](https://github.com/dagster-io/dagster/tree/master/examples/hacker_news):
+* [Specifying Ins and Outs](https://github.com/dagster-io/dagster/blob/master/examples/hacker_news/hacker_news/ops/comment_stories.py)
+* [Using resources](https://github.com/dagster-io/dagster/blob/master/examples/hacker_news/hacker_news/ops/download_items.py)
+* [Per-Output IO Manager](https://github.com/dagster-io/dagster/blob/master/examples/hacker_news/hacker_news/ops/user_top_recommended_stories.py)
