+++ docs/content/concepts/partitions-schedules-sensors/schedules.mdx
@@ -225,6 +225,12 @@ def test_configurable_job_schedule():
 
 If your `@schedule`-decorated function doesn't have a context parameter, you don't need to provide one when invoking it.
 
+## See it in action
+For more examples of schedules, check out the following in our [Hacker News example](https://github.com/dagster-io/dagster/tree/master/examples/hacker_news):
+* [Defining an hourly partitioned schedule](https://github.com/dagster-io/dagster/blob/master/examples/hacker_news/hacker_news/schedules/hourly_hn_download_schedule.py)
+* [Specifying the partition to a job](https://github.com/dagster-io/dagster/blob/master/examples/hacker_news/hacker_news/jobs/hacker_news_api_download.py)
+
+
 ## Troubleshooting
 
 Try these steps if you're trying to run a schedule and are running into problems.
